+++ lib/src/components/dropdown_menu/dropdown_menu_component.dart
@@ -103,6 +103,9 @@ class LiDropdownMenuComponent implements OnDestroy {
   String mobileMenuTitle = '';
 
   @Input()
+  String mobileCloseButtonAriaLabel = 'Fechar';
+
+  @Input()
   TemplateRef? mobileFooterTemplate;
 
   @Input()