+++ lib/src/components/offcanvas/offcanvas_component.dart
@@ -188,6 +188,9 @@ class LiOffcanvasComponent
   String titleText = '';
 
   @Input()
+  String closeButtonAriaLabel = 'Fechar';
+
+  @Input()
   String? offcanvasId;
 
   @Input()